Prevents unauthorized users from saving to db, creating new cart_item and deleting a cart_item.

+++ client/src/redux/cart/cart.utils.js
@@ -1,19 +1,21 @@
 import { firestore } from "../../utils/firebase.config";
 const cartItemColRef = firestore.collection("cart_items");
 
-export const addToCart = async (shoppingCart, item) => {
-  console.log(shoppingCart, "in add utils");
+export const addToCart = async (shoppingCart, item, isLoggedIn) => {
   const itemIndex = shoppingCart.findIndex(
     (cartItem) => cartItem.id === item.id
   );
   if (itemIndex === -1) {
-    const cartItemRef = cartItemColRef.doc();
-    const productRef = firestore.doc(`products/${item.id}`);
-    await cartItemRef.set({ productRef, quantity: 1 });
-    return [
-      ...shoppingCart,
-      { ...item, quantity: 1, cartItemId: cartItemRef.id }
-    ];
+    if (isLoggedIn) {
+      const cartItemRef = cartItemColRef.doc();
+      const productRef = firestore.doc(`products/${item.id}`);
+      await cartItemRef.set({ productRef, quantity: 1 });
+      return [
+        ...shoppingCart,
+        { ...item, quantity: 1, cartItemId: cartItemRef.id }
+      ];
+    }
+    return [...shoppingCart, { ...item, quantity: 1 }];
   }
   const prevQuantity = shoppingCart[itemIndex].quantity;
   shoppingCart[itemIndex] = {
@@ -23,9 +25,11 @@ export const addToCart = async (shoppingCart, item) => {
   return [...shoppingCart];
 };
 
-export const removeFromCart = async (shoppingCart, item) => {
-  const cartItemRef = firestore.doc(`cart_items/${item.cartItemId}`);
-  await cartItemRef.delete();
+export const removeFromCart = async (shoppingCart, item, isLoggedIn) => {
+  if (isLoggedIn) {
+    const cartItemRef = firestore.doc(`cart_items/${item.cartItemId}`);
+    await cartItemRef.delete();
+  }
   return shoppingCart.filter((cartItem) => cartItem.id !== item.id);
 };
 
@@ -41,7 +45,10 @@ export const changeItemQuantity = async (
     (cartItem) => cartItem.id === item.id
   );
   if (itemIndex !== -1) {
-    shoppingCart[itemIndex] = { ...item, quantity: newQuantity };
+    shoppingCart[itemIndex] = {
+      ...shoppingCart[itemIndex],
+      quantity: newQuantity
+    };
   }
   return [...shoppingCart];
 };
